Refactor tasks to separate file
